feat(money): migrate RRMS + config + order sizing to Decimal (Phase 2)
Second of three PRs implementing docs/DECIMAL_ENFORCEMENT_PLAN.md. The
money.py helpers shipped in Phase 1 (f4cd4a9) are now adopted across the
sizing and risk-gate code paths.

Zone boundaries honoured:
  Money zone (Decimal): config RRMS settings, rrms_engine, order_manager
    kill-switch + validation + open_positions, pretrade RRR check.
  Analysis zone (float): mwa_signal_generator — boundary cast added at
    `risk_amt = float(RRMS_CAPITAL * RRMS_RISK_PCT)` because risk itself
    is an ATR-derived float.
  SDK boundary (float): Kite + Angel place_order calls cast Decimal →
    float immediately before SDK dispatch (JSON payloads require float).

Changes:
- config.py: RRMS_CAPITAL and RRMS_RISK_PCT → Decimal via to_money().
  RRMS_MIN_RRR stays float (dimensionless ratio; never multiplied
  against money directly, only against ATR in analysis code).
- rrms_engine.py: RRMSResult money fields → Decimal; engine accepts
  Numeric at public API (int/float/str/Decimal), to_money() internally,
  round_tick() at result-build time so CDS currency pairs retain 4dp
  while equity/F&O round to paise.
- mwa_signal_generator.py: one-line boundary cast described above.
- pretrade_check.check_rrr: drop float(signal.rrr) cast — Signal.rrr is
  Decimal from the Numeric column; comparison against float 3.0 works.
- order_manager.py: capital + kill-switch + order_value + position_pct +
  trailing SL + SL-hit check + exit-strategy profit math all in Decimal.
  place_order accepts Numeric for price/stop_loss/target; converts at
  entry and casts back to float only for Kite/Angel SDK payloads.

Tests:
- test_rrms.py: 4 new tests assert Decimal field types, mixed Numeric
  input, CDS 4dp precision preservation, Decimal capital constructor.
- test_order_manager.py: trailing-SL short-position test asserts
  Decimal("581.40") directly — the Decimal path now produces the exact
  value that float multiplication truncated to 581.3999…

Verified: ruff clean on all touched files; 47 rrms+money tests pass;
82 focused order_manager/paper/pretrade/rrms/money tests pass (the 2
remaining order_manager failures are pre-existing stale "Kite not
connected" assertions from before the Angel broker merge and unrelated
to this change). Phase 3 will migrate signal_monitor + portfolio_risk
+ signal_cards.

+++ mcp_server/mwa_signal_generator.py
@@ -165,8 +165,10 @@ def generate_mwa_signals(
 
         rrr = (abs(target - entry)) / risk
 
-        # RRMS position sizing
-        risk_amt = settings.RRMS_CAPITAL * settings.RRMS_RISK_PCT
+        # RRMS position sizing — settings are Decimal (money zone); cast to
+        # float here because `risk` was computed from ATR/float OHLCV and this
+        # function is in the analysis zone.
+        risk_amt = float(settings.RRMS_CAPITAL * settings.RRMS_RISK_PCT)
         qty = floor(risk_amt / risk)
         if qty <= 0:
             qty = 1
